fixed null_file_sink_mt bug for ranks > 1; added tests to verify rank0 is the only rank with console output.  moved multi-log into its own file.
